requirements.txt
Django==1.3
-e hg+https://bitbucket.org/izi/django-admin-tools/@v0.4.0#egg=admin_tools
-e git://github.com/twidi/django-extended-choices.git#egg=extended_choices
django-extensions==0.6
-e hg+https://bitbucket.org/ubernostrum/django-profiles#egg=profiles
simplejson==2.1.6
south==0.7.3
-e git+git://github.com/digi604/django-smart-selects.git#egg=smart_selects
django-keyedcache==1.4-1
django-livesettings==1.4-7
django-smart-selects==1.0.4
sorl-thumbnail==11.09
PIL==1.1.7
django-floppyforms==0.4.7
feedparser==5.0.1
-e git+git://github.com/quinode/coop_cms.git#egg=coop_cms
-e git+git://github.com/quinode/djaloha.git#egg=djaloha
-e git+git://github.com/gabrielgrant/django-html-field.git#egg=html_field
djangoembed==0.1.1
django-taggit==0.9.3
django-taggit-autocomplete-modified==0.1.0b4
-e git+git://github.com/jonasgeiregat/django-templatetag-sugar.git#egg=templatetag_sugar
django-haystack==1.2.6
httplib2==0.7.2
python-dateutil==1.5